Create Methods payments

+++ src/Asaas/Models/Common/Discount.cs
@@ -4,7 +4,7 @@ namespace WP.Asaas.Sdk.Models.Common;
 
 public class Discount
 {
-    public decimal Value { get; init; }
+    public double Value { get; init; }
     public int DueDateLimitDays { get; init; }
     public DiscountType Type { get; init; }
 }
