+++ src/Asaas/Models/Invoice/MunicipalService.cs
@@ -4,5 +4,5 @@ public class MunicipalService
 {
     public string Id { get; init; }
     public string Description { get; init; }
-    public decimal Iss { get; init; }
+    public double Iss { get; init; }
 }
